Execute and fix tests in tests/unit/events
The tests were not executed as the file names did not start with "test_*" string.
Therefore, these were renamed. After that, there were few tests failing, this
is fixed in the commit as well.
