+++ tests/unit/events/test_copr.py
